style: rename response field
