+++ api-server/src/main/java/nexters/payout/apiserver/stock/application/dto/response/SectorRatioResponse.java
@@ -1,7 +1,7 @@
 package nexters.payout.apiserver.stock.application.dto.response;
 
 import nexters.payout.domain.stock.Sector;
-import nexters.payout.domain.stock.service.SectorAnalyzer.SectorInfo;
+import nexters.payout.domain.stock.service.SectorAnalysisService.SectorInfo;
 
 import java.util.List;
 import java.util.Map;
