+++ api-server/src/main/java/nexters/payout/apiserver/stock/application/dto/response/StockResponse.java
@@ -1,7 +1,6 @@
 package nexters.payout.apiserver.stock.application.dto.response;
 
 import nexters.payout.domain.dividend.Dividend;
-import nexters.payout.domain.stock.Sector;
 import nexters.payout.domain.stock.Stock;
 
 import java.util.UUID;
@@ -15,7 +14,7 @@ public record StockResponse(
         String industry,
         Double price,
         Integer volume,
-        Double dividend
+        Double dividendPerShare
 ) {
     public static StockResponse of(Stock stock, Dividend dividend) {
         return new StockResponse(
