+++ domain/build.gradle
@@ -25,6 +25,7 @@ bootJar.enabled = false
 jar.enabled = true
 
 dependencies {
+    implementation(project(":core"))
 
     // lombok
     compileOnly 'org.projectlombok:lombok'
