+++ domain/src/testFixtures/java/nexters/payout/domain/DividendFixture.java
@@ -16,6 +16,15 @@ public class DividendFixture {
         );
     }
 
+    public static Dividend createDividend(UUID stockId, Instant paymentDate) {
+        return Dividend.createDividend(
+                stockId,
+                12.21,
+                Instant.parse("2023-12-21T00:00:00Z"),
+                paymentDate,
+                Instant.parse("2023-12-22T00:00:00Z"));
+    }
+
     public static Dividend createDividend(UUID stockId) {
         return Dividend.createDividend(
                 stockId,
